requirements.txt
markdownmail==0.*
pyyaml==5.*
braceexpand==0.*